requirements.txt
country-converter==1.0.0
jishaku==2.5.1 # future versions are py3.10+
parsedatetime==2.6
tweepy==4.14.0
Pillow==9.5.0
markdownify==0.11.6
discord.py>=2.0.0,<3.0.0
git+https://github.com/Rapptz/discord-ext-menus@master
googletrans
aiohttp
asyncpg
humanize